исправлены ошибки
