+++ main.js
@@ -1,22 +1,15 @@
 import {logs} from './logs.js'
 import {data, getRandom} from './utils.js'
-import {HIT, ATTACK, player1, player2} from './variables.js'
+import {player1, player2} from './variables.js'
+import {playerAttack} from './playerAttack.js'
+import {enemyAttack} from './enemyAttack.js'
+import {createElement} from './createElement.js'
+import {showResult} from './showResult.js'
 
-const $arenas = document.querySelector('.arenas')
-const $randomBotton = document.querySelector('.button')
+export const $arenas = document.querySelector('.arenas')
+export const $randomBotton = document.querySelector('.button')
 const $chat = document.querySelector('.chat')
-
-const $formFight = document.querySelector('.control')
-
-
-const createElement = (tag, className) => {
-    const $tag = document.createElement(tag)
-    if (className) {
-        $tag.classList.add(className)
-    }
-    return $tag
-}
-
+export const $formFight = document.querySelector('.control')
 
 const createPlayer = (playerObj) => {
     const $player = createElement('div', 'player' + playerObj.player)
@@ -41,77 +34,11 @@ const createPlayer = (playerObj) => {
     return $player
 }
 
-const playerLose = (name) => {
-    const $loseTitle = createElement('div', 'loseTitle')
-
-    if (name) {
-        $loseTitle.innerText = name + ' wins'
-
-    } else {
-        $loseTitle.innerText = 'draw'
-    }
-
-    $arenas.appendChild(createReloadButton())
-    return $loseTitle
-}
-
-const createReloadButton = () => {
-    const $reloadWrap = createElement('div', 'reloadWrap')
-    const $button = createElement('button', 'button')
-    $button.innerHTML = "restart";
-
-    $reloadWrap.appendChild($button)
-    $reloadWrap.addEventListener('click', function () {
-        window.location.reload()
-    })
-    return $reloadWrap
-}
-
 $arenas.appendChild(createPlayer(player1))
 $arenas.appendChild(createPlayer(player2))
 
-const enemyAttack = () => {
-    const hit = ATTACK[getRandom(3) - 1]
-    const defence = ATTACK[getRandom(3) - 1]
 
-    return {
-        value: getRandom([HIT[hit]]),
-        hit,
-        defence
-    }
-}
-
-const playerAttack = () => {
-    const attack = {}
-    for (let item of $formFight) {
-        if (item.checked && item.name === 'hit') {
-            attack.value = getRandom(HIT[item.value])
-            attack.hit = item.value
-        }
-        if (item.checked && item.name === 'defence') {
-            attack.defence = item.value
-        }
-        item.checked = false;
-    }
-    return attack
-}
-
-const showResult = () => {
-    if (player1.hp === 0 && player1.hp < player2.hp) {
-        $arenas.appendChild(playerLose(player2.name))
-        generateLogs('end', player1, player2)
-    } else if (player2.hp === 0 && player2.hp < player1.hp) {
-        $arenas.appendChild(playerLose(player1.name))
-        generateLogs('end', player2, player1)
-    } else if (player1.hp === 0 && player2.hp === 0) {
-        $randomBotton.disabled = true
-        $arenas.appendChild(playerLose())
-        generateLogs('draw')
-    }
-
-}
-
-const generateLogs = (type, player1, player2) => {
+export const generateLogs = (type, player1, player2) => {
  
       const end = logs.end[1].replace('[playerWins]', player2.name).replace('[playerLose]', player1.name)
         let el
@@ -133,7 +60,8 @@ const generateLogs = (type, player1, player2) => {
             el = `<p>${data} ${logs[type][getRandom(logs[type].length -1)].replace('[playerKick]', player2.name).replace('[playerDefence]', player1.name)} - ${100 - player1.hp} [${player2.hp} / 100] <p>`;
             break;
         default: 
-        alert('ошибка')
+        console.log('ошибка');
+        break;
     }
    
     $chat.insertAdjacentHTML('afterbegin', el)
@@ -153,12 +81,14 @@ $formFight.addEventListener('submit', function (e) {
         player1.changeHP(enemy.value);
         player1.renderHP
         generateLogs('hit', player2, player1)
+        generateLogs('defence', player1, player2)
     }
 
     if (enemy.defence !== player.hit) {
         player2.changeHP(player.value);
         player2.renderHP()
         generateLogs('hit', player1, player2)
+        generateLogs('defence', player2, player1)
     }
 
     showResult()
